+++ generators/fourier_transform.hpp
@@ -21,11 +21,11 @@ namespace Texpainter::Generators
 		FourierTransform(FourierTransform const&) = delete;
 		FourierTransform& operator=(FourierTransform const&) = delete;
 
-		Model::BasicImage<std::complex<double>> operator()(Span2d<float const> vals_in);
+		Model::BasicImage<std::complex<double>> operator()(Span2d<double const> vals_in);
 
 		// This function should return a double, in case filtering makes the output range
 		// small
-		Model::BasicImage<float> operator()(Span2d<std::complex<double> const> vals_in);
+		Model::BasicImage<double> operator()(Span2d<std::complex<double> const> vals_in);
 
 		~FourierTransform();
